pyproject.toml
[tool.poetry]
name = "test-task"
version = "0.1.0"
description = ""
authors = ["Bohdan <AdamThomasMoran124@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
openai = "^1.59.7"
faiss-cpu = "^1.9.0.post1"
langchain = "^0.3.14"
pandas = "^2.2.3"
numpy = "^2.2.1"
sentence-transformers = "^3.3.1"
langchain-community = "^0.3.14"
langchain-openai = "^0.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
